+++ test/unit/git_class/unit_test_run.sh
@@ -73,6 +73,14 @@ echo "Unit test:  gitmerge_is_remote_branch"
 test/unit/git_class/gitmerge_is_remote_branch.py
 
 echo ""
+echo "Unit test:  gitmerge_get_dirty"
+test/unit/git_class/gitmerge_get_dirty.py
+
+echo ""
+echo "Unit test:  gitmerge_get_untracked"
+test/unit/git_class/gitmerge_get_untracked.py
+
+echo ""
 echo "Unit test:  gitclass_init"
 test/unit/git_class/gitclass_init.py