+++ CHANGELOG.md
@@ -4,13 +4,43 @@ All notable changes to this project will be documented in this file.
 The format is based on "Keep a Changelog".  This project adheres to Semantic Versioning.
 
 
+## [0.0.4] - 2019-04-29
+### Added
+- post_body:  Append default post-header to mail body.
+- quarantine_files:  Copy files out of Git repo into a quarantine directory.
+- quarantine:  Get dirty and untracked files and quarantine them.
+
+### Fixed
+- process_project:  Corrected call to git_fetch method.
+- prepare_mail:  Corrected logic in 'if' statement.
+- prepare_mail:  Fixed problem with mutable default arguments issue.
+- post_process:  Fixed problem with mutable default arguments issue.
+- merge:  Fixed problem with mutable default arguments issue.
+- run_program:  Fixed problem with mutable default arguments issue.
+- send_mail:  Added newline for each line in email.
+- post_check:  Added branch name as arg to is_commits_ahead and is_commits_behind calls.
+
+### Changed
+- Added function name to log entries.
+- main:  Set the repo name for -r option if not passed to program using -p option setting.
+- merge:  Added quarantine function call to quarantine new/modified files.
+- merge:  Added arguments to process_dirty and process_untracked method calls.
+- prepare_mail:  Replaced post body mail lines with call to post_body function.
+- post_process:  Added datetime to destination directory for the archiving of the repo.
+- merge:  Added datetime to destination directory for the archiving of the repo.
+- merge_project:  Added additional logging error statements.
+- process_project:  Added additional logging error statements.
+
+
 ## [0.0.3] - 2019-04-03
 ### Added
+- prepare_mail:  Build body of email message with a set of standard lines.
 - commits_diff:  Returns count difference between two branches.
 - is_commits_ahead:  Returns count on local branch ahead of remote branch.
 - is_commits_behind:  Returns count on local branch behind remote branch.
 
 ### Changed
+- Modified the program to use the git_class module, which the program will interface with Git through.
 - merge:  Added check to see if local and remote branches are in sync.